chore(smoke_test): Regenerate

+++ packages/aws_sdk/smoke_test/lib/src/sdk/src/ec2/model/delete_instance_event_window_request.dart
@@ -1,4 +1,5 @@
 // Generated with smithy-dart 0.3.1. DO NOT MODIFY.
+// ignore_for_file: avoid_unused_constructor_parameters,deprecated_member_use_from_same_package,non_constant_identifier_names
 
 library smoke_test.ec2.model.delete_instance_event_window_request; // ignore_for_file: no_leading_underscores_for_library_prefixes
 
@@ -115,28 +116,28 @@ class DeleteInstanceEventWindowRequestEc2QuerySerializer
     final result = DeleteInstanceEventWindowRequestBuilder();
     final iterator = serialized.iterator;
     while (iterator.moveNext()) {
-      final key = iterator.current;
+      final key = iterator.current as String;
       iterator.moveNext();
       final value = iterator.current;
-      switch (key as String) {
+      if (value == null) {
+        continue;
+      }
+      switch (key) {
         case 'DryRun':
           result.dryRun = (serializers.deserialize(
-            value!,
+            value,
             specifiedType: const FullType(bool),
           ) as bool);
-          break;
         case 'ForceDelete':
           result.forceDelete = (serializers.deserialize(
-            value!,
+            value,
             specifiedType: const FullType(bool),
           ) as bool);
-          break;
         case 'InstanceEventWindowId':
           result.instanceEventWindowId = (serializers.deserialize(
-            value!,
+            value,
             specifiedType: const FullType(String),
           ) as String);
-          break;
       }
     }
 
@@ -146,34 +147,38 @@ class DeleteInstanceEventWindowRequestEc2QuerySerializer
   @override
   Iterable<Object?> serialize(
     Serializers serializers,
-    Object? object, {
+    DeleteInstanceEventWindowRequest object, {
     FullType specifiedType = FullType.unspecified,
   }) {
-    final payload = (object as DeleteInstanceEventWindowRequest);
-    final result = <Object?>[
+    final result$ = <Object?>[
       const _i1.XmlElementName(
         'DeleteInstanceEventWindowRequestResponse',
         _i1.XmlNamespace('http://ec2.amazonaws.com/doc/2016-11-15'),
       )
     ];
-    result
+    final DeleteInstanceEventWindowRequest(
+      :dryRun,
+      :forceDelete,
+      :instanceEventWindowId
+    ) = object;
+    result$
       ..add(const _i1.XmlElementName('DryRun'))
       ..add(serializers.serialize(
-        payload.dryRun,
+        dryRun,
         specifiedType: const FullType(bool),
       ));
-    result
+    result$
       ..add(const _i1.XmlElementName('ForceDelete'))
       ..add(serializers.serialize(
-        payload.forceDelete,
+        forceDelete,
         specifiedType: const FullType(bool),
       ));
-    result
+    result$
       ..add(const _i1.XmlElementName('InstanceEventWindowId'))
       ..add(serializers.serialize(
-        payload.instanceEventWindowId,
+        instanceEventWindowId,
         specifiedType: const FullType(String),
       ));
-    return result;
+    return result$;
   }
 }
